Fix integer overflow in collision
DAmn spent so fucking many hours and evening finding why are those
cursed podes randomly turns out from the checkpoint with no confidence
it is a bug indeed and not just disability of 1-step random search to
make a good decision with this type of reward function. Finally noticed
that its a checkpoint passment detection that goes wrong and on the second
try printing everything in collision function found that got x*x < 0.
Really, 1e5 ** 2 is more than int32 can handle.